% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_versions.R
\name{available_ct_versions}
\alias{available_ct_versions}
\title{List available CT versions}
\usage{
available_ct_versions(type = c("sdtm", "adam"))
}
\arguments{
\item{type}{One of \code{"sdtm"} or \code{"adam"}.}
}
\value{
Character vector of version date strings, newest first,
  e.g. \code{c("2026-03-27", "2025-09-27", ...)}.
}
\description{
List available CT versions
}
\examples{
available_ct_versions("sdtm")
available_ct_versions("adam")
}
