% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ct_adam}
\alias{ct_adam}
\title{ADaM Controlled Terminology}
\format{
A data frame with the same columns as \code{\link{ct_sdtm}}.
}
\source{
\url{https://evs.nci.nih.gov/ftp1/CDISC/ADaM/}
}
\usage{
ct_adam
}
\description{
Versioned CDISC ADaM Controlled Terminology from the NCI EVS FTP site.
Same validity-date design as \code{\link{ct_sdtm}}.
}
\keyword{datasets}
