% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{List all available datasets in cdiscdata}
\usage{
list_datasets()
}
\value{
A data frame with one row per dataset describing its type, version
  range, and latest available version.
}
\description{
List all available datasets in cdiscdata
}
\examples{
list_datasets()
}
