% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{schema_path}
\alias{schema_path}
\title{Get path to a Define-XML XSD schema file}
\usage{
schema_path(define_version = c("2.1", "2.0"))
}
\arguments{
\item{define_version}{One of \code{"2.1"} or \code{"2.0"}.}
}
\value{
Full file path to the primary XSD file.
}
\description{
Get path to a Define-XML XSD schema file
}
\examples{
schema_path("2.1")
schema_path("2.0")
}
