% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{stylesheet_path}
\alias{stylesheet_path}
\title{Get path to a Define-XML XSLT stylesheet}
\usage{
stylesheet_path(define_version = c("2.1", "2.0"))
}
\arguments{
\item{define_version}{One of \code{"2.1"} or \code{"2.0"}.}
}
\value{
Full file path to the XSL stylesheet.
}
\description{
Get path to a Define-XML XSLT stylesheet
}
\examples{
stylesheet_path("2.1")
stylesheet_path("2.0")
}
